requirements.txt
numpy==1.24.3
Pillow==10.0.0
scikit-learn==1.3.0
geotessera